:truck: Move and rename test folders
for better readability
