--- jogakbo/src/test/java/com/noyes/jogakbo/auth/JwtCheck.java
@@ -1,5 +1,0 @@
-package com.noyes.jogakbo.auth;
-
-public class JwtCheck {
-
-}
